rename tools agent
