+++ azure-java-migration-copilot-client/src/main/java/com/azure/migration/java/copilot/service/common/Configure.java
@@ -1,0 +1,17 @@
+package com.azure.migration.java.copilot.service.common;
+
+import dev.langchain4j.model.chat.ChatLanguageModel;
+import dev.langchain4j.service.AiServices;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class Configure {
+
+    @Bean
+    ToolsAgent resourceToolsAgent(ChatLanguageModel chatLanguageModel) {
+        return AiServices.builder(ToolsAgent.class)
+                .chatLanguageModel(chatLanguageModel)
+                .build();
+    }
+}
