+++ azure-java-migration-copilot-client/src/main/java/com/azure/migration/java/copilot/service/common/ToolsAgent.java
@@ -1,10 +1,10 @@
-package com.azure.migration.java.copilot.service.resource;
+package com.azure.migration.java.copilot.service.common;
 
 import dev.langchain4j.service.SystemMessage;
 import dev.langchain4j.service.UserMessage;
 import dev.langchain4j.service.V;
 
-public interface ResourceConfigureNoChatAgent {
+public interface ToolsAgent {
 
     @SystemMessage(fromResource = "/prompts/resource/resource-abstract.txt")
     String abstractInfo(@V("schema") String schema,@V("orignalData") String orignalData,@UserMessage String userMessage);
