+++ azure-java-migration-copilot-client/src/main/java/com/azure/migration/java/copilot/service/resource/ResourceConfigure.java
@@ -19,11 +19,4 @@ public class ResourceConfigure {
                 .contentRetriever(contentRetriever)
                 .build();
     }
-
-    @Bean
-    ResourceConfigureNoChatAgent resourceConfigureNoChatAgent(ChatLanguageModel chatLanguageModel) {
-        return AiServices.builder(ResourceConfigureNoChatAgent.class)
-                .chatLanguageModel(chatLanguageModel)
-                .build();
-    }
 }
